requirements.txt
# python3.12.5
tensorflow==2.17
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
opencv-python==4.10.0.84
Pillow==10.4.0
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
Flask==3.0.3